chore: rename dagreCompound layout with Camel-Case (#70)
* chore: rename dagreCompound layout with Camel-Case

* fix: remove any definition in dagre compound layout
